working solution added

+++ angular-10-firestore-crud-master/src/app/components/tutorial-details/tutorial-details.component.ts
@@ -1,6 +1,6 @@
 import { Component, OnInit, Input, OnChanges, Output, EventEmitter } from '@angular/core';
+import Tutorial from 'src/app/models/tutorial.model';
 import { TutorialService } from 'src/app/services/tutorial.service';
-import Tutorial from 'src/app/models/tutorial';
 
 @Component({
   selector: 'app-tutorial-details',
@@ -9,9 +9,13 @@ import Tutorial from 'src/app/models/tutorial';
 })
 export class TutorialDetailsComponent implements OnInit, OnChanges {
 
-  @Input() tutorial: Tutorial;
+  @Input() tutorial?: Tutorial;
   @Output() refreshList: EventEmitter<any> = new EventEmitter();
-  currentTutorial: Tutorial = null;
+  currentTutorial: Tutorial = {
+    title: '',
+    description: '',
+    published: false
+  };
   message = '';
 
   constructor(private tutorialService: TutorialService) { }
@@ -25,13 +29,15 @@ export class TutorialDetailsComponent implements OnInit, OnChanges {
     this.currentTutorial = { ...this.tutorial };
   }
 
-  updatePublished(status): void {
-    this.tutorialService.update(this.currentTutorial.id, { published: status })
+  updatePublished(status: boolean): void {
+    if (this.currentTutorial.id) {
+      this.tutorialService.update(this.currentTutorial.id, { published: status })
       .then(() => {
         this.currentTutorial.published = status;
         this.message = 'The status was updated successfully!';
       })
       .catch(err => console.log(err));
+    }
   }
 
   updateTutorial(): void {
@@ -40,18 +46,22 @@ export class TutorialDetailsComponent implements OnInit, OnChanges {
       description: this.currentTutorial.description
     };
 
-    this.tutorialService.update(this.currentTutorial.id, data)
-      .then(() => this.message = 'The tutorial was updated successfully!')
-      .catch(err => console.log(err));
+    if (this.currentTutorial.id) {
+      this.tutorialService.update(this.currentTutorial.id, data)
+        .then(() => this.message = 'The tutorial was updated successfully!')
+        .catch(err => console.log(err));
+    }
   }
 
   deleteTutorial(): void {
-    this.tutorialService.delete(this.currentTutorial.id)
-      .then(() => {
-        this.refreshList.emit();
-        this.message = 'The tutorial was updated successfully!';
-      })
-      .catch(err => console.log(err));
+    if (this.currentTutorial.id) {
+      this.tutorialService.delete(this.currentTutorial.id)
+        .then(() => {
+          this.refreshList.emit();
+          this.message = 'The tutorial was updated successfully!';
+        })
+        .catch(err => console.log(err));
+    }
   }
 
 }
